backend evenement

+++ src/main/java/miage/groupe2/reseausocial/Controller/PostController.java
@@ -46,7 +46,7 @@ public class PostController {
 
     @GetMapping("/{id}")
     public String postPersonne(
-            @PathVariable long id,
+            @PathVariable Integer id,
             Model model
     ) {
         Utilisateur user = utilisateurRepository.findByIdUti(id);
